Refactor: Move video embed components from Note to Post directory
